move EnsureAddonConfig to common pkg
