+++ internal/metrics/resource/resource_test.go
@@ -170,159 +170,6 @@ func TestReconcileAgent(t *testing.T) {
 	assert.Equal(t, foundAgent.Spec.ScrapeConfigNamespaceSelector, &metav1.LabelSelector{})
 }
 
-func TestEnsureAddonConfig(t *testing.T) {
-	platformConfig := addonv1alpha1.AddOnConfig{
-		ConfigReferent: addonv1alpha1.ConfigReferent{
-			Namespace: "ns",
-			Name:      "platform-agent",
-		},
-		ConfigGroupResource: addonv1alpha1.ConfigGroupResource{
-			Group:    prometheusalpha1.SchemeGroupVersion.Group,
-			Resource: prometheusalpha1.PrometheusAgentName,
-		},
-	}
-	uwlConfig := addonv1alpha1.AddOnConfig{
-		ConfigReferent: addonv1alpha1.ConfigReferent{
-			Namespace: "ns",
-			Name:      "uwl-agent",
-		},
-		ConfigGroupResource: addonv1alpha1.ConfigGroupResource{
-			Group:    prometheusalpha1.SchemeGroupVersion.Group,
-			Resource: prometheusalpha1.PrometheusAgentName,
-		},
-	}
-
-	placementRefA := addonv1alpha1.PlacementRef{
-		Namespace: "ns",
-		Name:      "a",
-	}
-	placementRefB := addonv1alpha1.PlacementRef{
-		Namespace: "ns",
-		Name:      "b",
-	}
-
-	testCases := []struct {
-		name                    string
-		initialPlacements       []addonv1alpha1.PlacementStrategy
-		inputConfigs            []defaultConfig
-		expectUpdateCalls       int
-		expectedFinalPlacements []addonv1alpha1.PlacementStrategy
-	}{
-		{
-			name: "no configs to add",
-			initialPlacements: []addonv1alpha1.PlacementStrategy{
-				{
-					Configs:      []addonv1alpha1.AddOnConfig{platformConfig},
-					PlacementRef: placementRefA,
-				},
-			},
-			inputConfigs: []defaultConfig{
-				{placementRef: placementRefA, config: platformConfig},
-			},
-			expectUpdateCalls: 0,
-			expectedFinalPlacements: []addonv1alpha1.PlacementStrategy{
-				{
-					Configs:      []addonv1alpha1.AddOnConfig{platformConfig},
-					PlacementRef: placementRefA,
-				},
-			},
-		},
-		{
-			name: "one configs to add in one placement",
-			initialPlacements: []addonv1alpha1.PlacementStrategy{
-				{
-					Configs:      []addonv1alpha1.AddOnConfig{platformConfig},
-					PlacementRef: placementRefA,
-				},
-				{
-					PlacementRef: placementRefB,
-				},
-			},
-			inputConfigs: []defaultConfig{
-				{placementRef: placementRefA, config: platformConfig},
-				{placementRef: placementRefB, config: platformConfig},
-			},
-			expectUpdateCalls: 1,
-			expectedFinalPlacements: []addonv1alpha1.PlacementStrategy{
-				{
-					Configs:      []addonv1alpha1.AddOnConfig{platformConfig},
-					PlacementRef: placementRefA,
-				},
-				{
-					Configs:      []addonv1alpha1.AddOnConfig{platformConfig},
-					PlacementRef: placementRefB,
-				},
-			},
-		},
-		{
-			name:              "no placement",
-			initialPlacements: []addonv1alpha1.PlacementStrategy{},
-			inputConfigs: []defaultConfig{
-				{placementRef: placementRefA, config: platformConfig},
-				{placementRef: placementRefB, config: platformConfig},
-			},
-			expectUpdateCalls:       0,
-			expectedFinalPlacements: []addonv1alpha1.PlacementStrategy{},
-		},
-		{
-			name: "multiple configs to add in multiple placements",
-			initialPlacements: []addonv1alpha1.PlacementStrategy{
-				{
-					Configs:      []addonv1alpha1.AddOnConfig{platformConfig},
-					PlacementRef: placementRefA,
-				},
-				{
-					PlacementRef: placementRefB,
-				},
-			},
-			inputConfigs: []defaultConfig{
-				{placementRef: placementRefA, config: platformConfig},
-				{placementRef: placementRefB, config: platformConfig},
-				{placementRef: placementRefA, config: uwlConfig},
-				{placementRef: placementRefB, config: uwlConfig},
-			},
-			expectUpdateCalls: 1,
-			expectedFinalPlacements: []addonv1alpha1.PlacementStrategy{
-				{
-					Configs:      []addonv1alpha1.AddOnConfig{platformConfig, uwlConfig},
-					PlacementRef: placementRefA,
-				},
-				{
-					Configs:      []addonv1alpha1.AddOnConfig{platformConfig, uwlConfig},
-					PlacementRef: placementRefB,
-				},
-			},
-		},
-	}
-
-	for _, tt := range testCases {
-		t.Run(tt.name, func(t *testing.T) {
-			cmao := newCMAO(tt.initialPlacements...)
-			updateCalls := 0
-			fakeClient := fake.NewClientBuilder().WithInterceptorFuncs(interceptor.Funcs{
-				Update: func(ctx context.Context, clientww client.WithWatch, obj client.Object, opts ...client.UpdateOption) error {
-					updateCalls++
-					return clientww.Update(ctx, obj, opts...)
-				},
-			}).WithScheme(newTestScheme()).WithObjects(cmao).Build()
-			d := DefaultStackResources{
-				Client: fakeClient,
-				CMAO:   cmao,
-				Logger: klog.Background(),
-			}
-
-			err := d.ensureAddonConfig(context.Background(), tt.inputConfigs)
-			assert.NoError(t, err)
-			assert.Equal(t, tt.expectUpdateCalls, updateCalls)
-
-			finalCmao := &addonv1alpha1.ClusterManagementAddOn{}
-			err = fakeClient.Get(context.Background(), client.ObjectKeyFromObject(cmao), finalCmao)
-			assert.NoError(t, err)
-			assert.ElementsMatch(t, tt.expectedFinalPlacements, finalCmao.Spec.InstallStrategy.Placements)
-		})
-	}
-}
-
 func TestReconcile(t *testing.T) {
 	// variables: placements, agents, addonOptions
 	placementRefA := addonv1alpha1.PlacementRef{
